Function ed_join::qgram::generate_inverted_index

pub fn generate_inverted_index(
    doc_x: &PathBuf, 
    doc_y: &PathBuf, 
    q: usize
) -> Result<InvertedIndex, Error>

This function reads an entire file into a string, count q-grams by parallel iterators, and returns a hashmap where the keys are q-gram tokens, and values are a vector of line-position pair.

NOTE: If the input file is larger, then simply divide the input to small slices in pre-processing. This would be implemented in the future.

Args

doc_x and doc_y: Path, absolute or relative, to documents to be processed.

file_x_len and file_y_len: Number of lines for each file. This is used

q: A tuning parameter used to generate the q-grams.

Returns

When succesful, returns a B-Tree map, where keys are numbers of occurences of all token, i.e. a q-gram, and values are vectors of tokens that have that number of occurences.