Function ed_join::verification::compare_qgrams

fn compare_qgrams(
    x: &PosQGramArray, 
    y: &PosQGramArray, 
    invert: &InvertedIndex, 
    tau: usize
) -> (PosQGramArray, usize)

Given two sorted q-gram arrays, in increasing order of location, find the set of loosely mismatching q-grams and the number of strictly mismatching q-grams.

Parameters

x and y: PosQGramArrays as source and target for matching.

invert: The inverted index.

tau: A positive integer as the tuning parameter for threshold for matching.

Return

A set of loosely mismatching q-grams from x to y.

The number of strictly mismatching q-grams from x to y.