Function ed_join::verification::content_filter

fn content_filter(
    from: &str, 
    to: &str, 
    mismatch: PosQGramArray, 
    suffix_sum: Vec<(usize, usize)>, 
    q: usize, 
    tau: usize
) -> Option<usize>

Content-based mismatch filtering by combining L1-distance and minimum edit errors in the suffix to the probing window.

Parameters

from and to: (Sub-)String that is under probing window.

mismatch: A PosQGramArray with loosely mismatching q-grams from s to t.

suffix_sum: A condensed suffix sum array.

q: A positive integer as the tuning parameter for length of q-grams.

tau: A positive integer as the tuning parameter for threshold for matching.

Return

A lower bound of the edit distance from s to t.